+++ spotinst_sdk2/clients/elastigroup/__init__.py
@@ -86,8 +86,6 @@ class ElastigroupAwsClient(Client):
 
         body_json = json.dumps(formatted_group_update_dict)
 
-        self.print_output(body_json)
-
         group_response = self.send_put_with_params(
             body=body_json,
             url=self.__base_elastigroup_url + "/" + group_id,