+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.1.32] - 2023-04-03
+### Fixed
+- Fixed logging issue for response object.
+
+## [2.1.31] - 2023-04-03
+### Added
+- Add support for AWS Ocean API: `fetch_rightsizing_recommendations`
+- Add request and response logging in client at DEBUG level.
+### Fixed
+- Fixed issue of adding multiple streamhandlers for each client.
+
 ## [2.1.30] - 2023-03-19
 ### Added
 - Add `immediate_o_d_recover_threshold` field in `Strategy` model in spotinst_sdk2.models.elastigroup.aws